third_party/WebKit/LayoutTests/http/tests/htmlimports/import-script-block-crossorigin-dynamic.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<link id="target" rel="import" href="http://localhost:8000/htmlimports/resources/cors-having-crossorigin-scripts.cgi">
</head>
<body>
<script>
description("This test ensures that crossorigin-marked script elements are blocked properly");
window.jsTestIsAsync = true;

function testAndDone()
{
    shouldBe("document.externalScriptHasRun", "undefined");
    shouldBe("document.corsExternalScriptHasRun", "undefined");
    shouldBeTrue("document.corsExternalScriptForLocalhostHasRun");
    finishJSTest();
}

// This should be called run-check.js
function check()
{
    testAndDone();
}

</script>
</body>
</html>
