third_party/WebKit/LayoutTests/http/tests/htmlimports/import-script-block-crossorigin-dynamic-expected.txt

CONSOLE ERROR: Access to Script at 'http://127.0.0.1:8000/htmlimports/resources/external-script.js' from origin 'http://localhost:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://localhost:8000' is therefore not allowed access.
CONSOLE ERROR: Access to Script at 'http://127.0.0.1:8000/htmlimports/resources/cors-js.cgi' from origin 'http://localhost:8000' has been blocked by CORS policy: The 'Access-Control-Allow-Origin' header has a value 'http://127.0.0.1:8000' that is not equal to the supplied origin. Origin 'http://localhost:8000' is therefore not allowed access.
This test ensures that crossorigin-marked script elements are blocked properly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.externalScriptHasRun is undefined
PASS document.corsExternalScriptHasRun is undefined
PASS document.corsExternalScriptForLocalhostHasRun is true
PASS successfullyParsed is true

TEST COMPLETE
